[workloadmeta] Skip events for unchanged entities
This is a minor performance improvement. Collectors generate identical
entities very often, when the changes to them are not relevant to
workloadmeta (or, in the case of pulling collectors, on every pull).
This has some minor overhead internally, generating deep-copies of
unchanged objects for bookkeeping, but has a somewhat bigger impact
externally, as every subscriber gets notified of basically no-op events.
This has an *even bigger* impact on the remote tagger, as those no-op
events get serialized, go over the "network" (locally, but still), and
get de-serialized on the other end. So this patch operates under the
assumption that a deep-equal check is much cheaper than all that.